+++ packages/core/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from "./analysis";
+export * from "./rules";
 export * from "./segment";
 export * from "./trace";